backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
openai   # optional if using OpenAI
python-dotenv
httpx